\name{ABC.Interact.dat}
\docType{data}
\alias{ABC.Interact.dat}
\title{Randomly generated set of values indexed by three factors}
\description{
This data set has randomly generated values of the response variable MOE
(Measure Of Effectiveness) which is indexed by the two-level factors A, B and C.
}
\usage{data(ABC.Interact.dat)}
\format{A data.frame containing 8 observations of 4 variables.}
\source{Generated by Chris Brien}
\keyword{datasets}