\name{BIBDWheat.dat}
\docType{data}
\alias{BIBDWheat.dat}
\title{Data for a balanced incomplete block experiment}
\description{
The data set comes from Joshi (1987) and is the data from an experiment to investigate 
six varieties of wheat that employs a balanced incomplete block design (BIBD) with ten blocks, 
each consisting of three plots. For more details see the vignette accessed via \code{vignette("DesignNotes", package="dae")}.
}
\usage{data(BIBDWheat.dat)}
\format{A data.frame containing 30 observations of 4 variables.}
\source{
Joshi, D. D. (1987) \emph{Linear Estimation and Design of Experiments}. Wiley Eastern, New Delhi.}
\keyword{datasets}