\name{Cabinet1.des}
\docType{data}
\alias{Cabinet1.des}
\title{A design for one of the growth cabinets in an experiment with 50 lines and 4 harvests}
\description{
The systematic design for a lattice square for 49 treatments consisting of four 7 x 7 squares. For more details see the vignette \cite{daeDesignNotes.pdf}.
}
\usage{data(Cabinet1.des)}
\format{A data.frame containing 160 observations of 15 variables.}
\keyword{datasets}