\name{Casuarina.dat}
\docType{data}
\alias{Casuarina.dat}
\title{Data for an experiment with rows and columns from Williams (2002)}
\description{
Williams (2002, p.144) provides an example of a resolved, Latinized, row-column design with four rectangles (blocks) each of six rows by ten columns. The experiment investigated differences between 60 provenances of a species of Casuarina tree, these provenances coming from 18 countries; the trees were inoculated prior to planting at two different times, time of inoculation being assigned to the four replicates so that each occurred in two replicates. At 30 months, diameter at breast height (Dbh) was measured. For more details see the vignette \cite{daeDesignNotes.pdf}.
}
\usage{data(Casuarina.dat)}
\format{A data.frame containing 240 observations of 7 variables.}
\source{Williams, E. R., Matheson, A. C. and Harwood, C. E. (2002) \emph{Experimental design and analysis for tree improvement}. 2nd edition. CSIRO, Melbourne, Australia.}
\keyword{datasets}