\name{dae-deprecated}
\alias{dae-deprecated}
\alias{proj2.decomp}
\alias{proj2.ops}
\title{Deprecated Functions in Package dae}
\description{These functions have been renamed and deprecated in \code{dae}.}
\usage{proj2.decomp(...)
proj2.ops(...)}
\arguments{
  \item{...}{absorbs arguments passed from the old functions of the style foo.bar().}
}
\author{Chris Brien}
