\name{daeTips}
\alias{daeTips}
\title{The intermittent, randomly-presented, startup tips.}
\description{The intermittent, randomly-presented, startup tips.}

\section{Startup tips}{
Need help? Enter help(package = 'dae') and click on 'User guides, package vignettes and other docs'.

Need help? The manual is in the doc subdirectory of the package's install directory.

Find out what has changed in dae: enter news(package = 'dae').

Need help to produce randomized designs? Enter help(package = 'dae') and look for the package vignettes. 

Need help to do the canonical analyis of a design? go to http://chris.brien.name/rpackages. 

Use suppressPackageStartupMessages() to eliminate all package startup messages.

To see all the intermittent, randomly-presented, startup tips enter ?daeTips.

For versions between CRAN releases (and more) go to \url{http://chris.brien.name/rpackages}.
}

\author{Chris Brien}

\keyword{asreml}