% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deloitte_cols.R
\docType{data}
\name{deloitte_colours}
\alias{deloitte_colours}
\title{deloitte_cols}
\format{
An object of class \code{character} of length 47.
}
\usage{
deloitte_colours
}
\description{
Function to select from a list of deloitte colours
}
\examples{

}
\keyword{datasets}
