% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deloitte_pal.R
\docType{data}
\name{deloitte_palettes}
\alias{deloitte_palettes}
\title{deloitte_pal}
\format{
An object of class \code{list} of length 5.
}
\usage{
deloitte_palettes
}
\arguments{
\item{gradient}{}
}
\description{
Deloitte palette selector
}
\examples{

}
\keyword{datasets}
