\name{Exp249.mplot.des}
\docType{data}
\alias{Exp249.mplot.des}
\title{Systematic, main-plot design for an experiment to be run in a greenhouse}
\description{
In this main-plot design, there are 24 lanes by 11 Positions, the lanes being blocked into 6 Zones of 4 lanes. The design for the main plots is for assigning 75 wheat lines, of which 73 are Nested Association Mapping (NAM) wheat lines and the other two are two check lines, Scout and Gladius. A row and column design was generated with \code{DiGGer} (Coombes, 2009). For more details see the vignette \cite{daeDesignNotes.pdf}.
}
\usage{data(Exp249.mplot.des)}
\format{A data.frame containing 264 observations of 3 variables.}
\source{Coombes, N. E. (2009) code{Digger}\emph{: design search tool in R}. {URL:} {\url{http://nswdpibiom.org/austatgen/software/}},
                 (accessed June 3, 2015).}
\keyword{datasets}