\name{Fac4Proc.dat}
\docType{data}
\alias{Fac4Proc.dat}
\title{Data for a 2^4 factorial experiment}
\description{
The data set come from an unreplicated \eqn{2^4}{2^4} factorial experiment to 
investigate a chemical process. The response variable is the Conversion 
percentage (Conv) and this is indexed by the 4 two-level factors Catal, Temp, 
Press and Conc, with levels \dQuote{-} and \dQuote{+}. The data is aranged in 
Yates order. Also included is the 16-level factor Runs which gives the order 
in which the combinations of the two-level factors were run.
}
\usage{data(Fac4Proc.dat)}
\format{A data.frame containing 16 observations of 6 variables.}
\source{
Table 10.6 of Box, Hunter and Hunter (1978) \emph{Statistics for Experimenters}. 
New York, Wiley.}
\keyword{datasets}