\name{get.daeRNGkind}
\alias{get.daeRNGkind}
\title{Gets the value of daeRNGkind for the package dae from the daeEnv environment}
\description{A function that gets the character value of \code{daeRNGkind} from the 
\code{daeEnv} environment. The value specifies the name of the Random Number Generator to use 
in \code{dae}.}
\usage{get.daeRNGkind()}
\value{The \code{character} value of \code{daeRNGkind}.}
\author{Chris Brien}
\seealso{\code{\link{set.daeRNGkind}}.}
\examples{
## get daeRNGkind.
get.daeRNGkind()
}
\keyword{manip}
