\name{get.daeTolerance}
\alias{get.daeTolerance}
\title{Gets the value of {daeTolerance} for the package dae}
\description{A function that gets the value such that, in \pkg{dae} functions, 
values less than it are considered to be zero.}
\usage{get.daeTolerance()}
\value{The value of {daeTolerance}.}
\author{Chris Brien}
\seealso{\code{\link{set.daeTolerance}}.}
\examples{
## get daeTolerance.
get.daeTolerance()
}
\keyword{manip}
\keyword{projector}