\name{LatticeSquare_t49.des}
\docType{data}
\alias{LatticeSquare_t49.des}
\title{A Lattice square design for 49 treatments}
\description{
The systematic design for a lattice square for 49 treatments consisting of four 7 x 7 squares. For more details see the vignette \cite{daeDesignNotes.pdf}.
}
\usage{data(LatticeSquare_t49.des)}
\format{A data.frame containing 196 observations of 4 variables.}
\source{Cochran and Cox (1957) \emph{Experimental Designs}. 2nd edn Wiley, New York.}
\keyword{datasets}