\name{McIntyreTMV.dat}
\docType{data}
\alias{McIntyreTMV.dat}
\title{The design and data from McIntyre's (1955) two-phase experiment}
\description{
McIntyre (1955) reports an investigation of the effect of four light intensities on the synthesis of tobacco mosaic virus in leaves of tobacco \emph{Nicotiana tabacum} var. Hickory Pryor. It is a two-phase experiment: the first phase is a treatment phase, in which the four light treatments are randomized to the tobacco leaves, and the second phase is an assay phase, in which the tobacco leaves are randomized to the half-leaves of assay plants. For more details see the vignette accessed via \code{vignette("DesignNotes", package="dae")}.}
\usage{data(McIntyreTMV.dat)}
\format{A data.frame containing 196 observations of 4 variables.}
\source{McIntyre, G. A. (1955) Design and Analysis of Two Phase Experiments. \emph{Biometrics}, \bold{11}, 324--334.}
\keyword{datasets}