% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_deloitte.R
\name{scale_colour_deloitte}
\alias{scale_colour_deloitte}
\title{scale_colour_deloitte}
\usage{
scale_colour_deloitte(
  palette = "2019",
  discrete = TRUE,
  reverse = FALSE,
  gradient = FALSE,
  ...
)
}
\arguments{
\item{palette}{ List of colours to use. \emph{Default palette is 2019}.}
\item{discrete}{To specify whether the data is discrete (TRUE) or continuous (FALSE). \emph{Default is TRUE}.}
\item{reverese}{Reverse the order of the colours in the chosen colour palette if TRUE. \emph{Default is no reverse ordering (FALSE)}.
}
\item{gradient}{Apply a diverging colour gradient based on the colours chose in palette if TRUE. Recommended to apply gradient if there are more variables than colours in the palette. \emph{Default is no gradient (FALSE)}.}
}
\description{
The \emph{scale_colour_deloitte} function allows the user to choose the colours of the lines in the chart.
}
\examples{
ggplot(mpg, aes(displ, hwy, colour = class)) +
  geom_point() +
  scale_colour_deloitte() +
  DAEChart


\strong{To create custom palettes, do the following}:
ggplot(mpg, aes(displ, hwy, colour = class)) +
  geom_point() +
  scale_colour_manual(values = list(deloitte_cols("green"), deloitte_cols("blue"))) +
  DAEChart
}
