% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_deloitte.R
\name{scale_fill_deloitte}
\alias{scale_fill_deloitte}
\title{scale_fill_deloitte}
\usage{
scale_fill_deloitte(
  palette = "2019",
  discrete = TRUE,
  reverse = FALSE,
  gradient = FALSE,
  ...
)
}
\arguments{
\item{gradient}{}
}
\description{
Fill scale constructor
}
\examples{

}
