\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,ANY-method}
\alias{show,classRepresentation-method}
\alias{show,genericFunction-method}
\alias{show,MethodDefinition-method}
\alias{show,MethodSelectionReport-method}
\alias{show,MethodWithNext-method}
\alias{show,ObjectsWithPackage-method}
\alias{show,oldClass-method}
\alias{show,projector-method}
\alias{show,signature-method}
\alias{show,traceable-method}
\title{Methods for Function show in Package dae}
\description{Methods for function \code{show} in Package \pkg{dae}}
\section{Methods}{
\describe{
 \item{\code{signature(object = "projector")}}{Prints the \code{matrix} and its degrees of freedom.}
}}
\seealso{\code{\link[=projector-class]{projector}} for further information about this class.}
\keyword{methods}
\keyword{array}
\keyword{projector}
