\name{SPLGrass.dat}
\docType{data}
\alias{SPLGrass.dat}
\title{Data for an experiment to investigate the effects of grazing patterns on
pasture composition}
\description{
The response variable is the percentage area covered by the principal grass 
(Main.Grass). The design for the experiment is a split-unit design. The main units 
are arranged in 3 Rows x 3 Columns. Each main unit is split into 
2 SubRows x 2 SubColumns.

The factor Period, with levels 3, 9 and 18 days, is assigned to the main units 
using a 3 x 3 Latin square. The two-level factors Spring and Summer are 
assigned to split-units using a criss-cross design within each main unit. 
The levels of each of Spring and Summer are two different grazing patterns in its season.
}
\usage{data(SPLGrass.dat)}
\format{A data.frame containing 36 observations of 8 variables.}
\source{
Example 14.1 from Mead, R. (1990). \emph{The Design of Experiments: 
Statistical Principles for Practical Application.} Cambridge, 
Cambridge University Press.}
\keyword{datasets}