\name{dae-deprecated}
\alias{dae-deprecated}
\alias{Ameasures}
\alias{blockboundary.plot}
\alias{design.plot}
\alias{proj2.decomp}
\alias{proj2.ops}
\alias{projs.canon}
\alias{projs.structure}
\title{Deprecated Functions in Package dae}
\description{These functions have been renamed and deprecated in \code{dae}.}
\usage{blockboundary.plot(...)
design.plot(...)
proj2.decomp(...)
proj2.ops(...)
projs.canon(...)
projs.structure(...)}
\arguments{
  \item{...}{absorbs arguments passed from the old functions of the style foo.bar().}
}
\author{Chris Brien}
