% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_deloitte.R
\name{scale_colour_deloitte}
\alias{scale_colour_deloitte}
\title{scale_colour_deloitte}
\usage{
scale_colour_deloitte(
  palette = "2019",
  discrete = TRUE,
  reverse = FALSE,
  gradient = FALSE,
  ...
)
}
\arguments{
\item{gradient}{}
}
\description{
Colour scale constructor
}
\examples{

}
