% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_deloitte.R
\name{scale_colour_deloitte}
\alias{scale_colour_deloitte}
\title{scale_colour_deloitte}
\usage{
scale_colour_deloitte(
  palette = "2019",
  discrete = TRUE,
  reverse = FALSE,
  gradient = FALSE,
  ...
)
}
\arguments{
\item{palette}{Default to 2019 colours}
\item{discrete}{Works with discrete variables}
\item{reverese}{Reverses the order of colours}
\item{gradient}{Create a gradient}
}
\description{
Colour scale constructor
}
\examples{

}
